+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Tarantool benchmark tool update (cluster bench):
+  * option --leader has been added - sest array of url's for leaders.
+  * option --replica has been added - sest array of url's for replicas.
+
 ## [2.12.4] - 2022-12-31
 
 ### Changed